# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SenAI Agentic CRM Intelligence Platform",
    "version": "1.0.0"
  },
  "paths": {
    "/api/ingest": {
      "post": {
        "summary": "Post Ingest",
        "operationId": "post_ingest_api_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/status/{job_id}": {
      "get": {
        "summary": "Get Status",
        "operationId": "get_status_api_status__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/stats": {
      "get": {
        "summary": "Dashboard Stats",
        "operationId": "dashboard_stats_dashboard_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/threads/{contact_email}": {
      "get": {
        "summary": "Get Threads",
        "operationId": "get_threads_threads__contact_email__get",
        "parameters": [
          {
            "name": "contact_email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contact Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/respond/{email_id}": {
      "post": {
        "summary": "Respond",
        "operationId": "respond_respond__email_id__post",
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Email Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drafts/{id}": {
      "patch": {
        "summary": "Edit Draft",
        "operationId": "edit_draft_drafts__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drafts/{id}/approve": {
      "post": {
        "summary": "Approve Draft",
        "operationId": "approve_draft_drafts__id__approve_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          },
          {
            "name": "approved_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "user",
              "title": "Approved By"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/sentiment-trend": {
      "get": {
        "summary": "Sentiment Trend",
        "operationId": "sentiment_trend_analytics_sentiment_trend_get",
        "parameters": [
          {
            "name": "sender",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sender"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/category-breakdown": {
      "get": {
        "summary": "Category Breakdown",
        "operationId": "category_breakdown_analytics_category_breakdown_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/rag/search": {
      "get": {
        "summary": "Rag Search",
        "operationId": "rag_search_rag_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/intelligence/reputation": {
      "get": {
        "summary": "Reputation",
        "operationId": "reputation_intelligence_reputation_get",
        "parameters": [
          {
            "name": "company",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ourplatform",
              "title": "Company"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/dry-run/{email_id}": {
      "post": {
        "summary": "Agent Dry Run",
        "operationId": "agent_dry_run_agent_dry_run__email_id__post",
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Email Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audit/{entity_type}/{entity_id}": {
      "get": {
        "summary": "Get Audit",
        "operationId": "get_audit_audit__entity_type___entity_id__get",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Type"
            }
          },
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contacts/{email}": {
      "get": {
        "summary": "Contact Profile",
        "operationId": "contact_profile_contacts__email__get",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contacts/{email}/status": {
      "patch": {
        "summary": "Update Contact Status",
        "operationId": "update_contact_status_contacts__email__status_patch",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/seed/knowledge-base": {
      "post": {
        "summary": "Seed Kb",
        "operationId": "seed_kb_seed_knowledge_base_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ContactOut": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "account_value": {
            "type": "number",
            "title": "Account Value"
          },
          "churn_risk_score": {
            "type": "number",
            "title": "Churn Risk Score"
          },
          "open_threads": {
            "type": "integer",
            "title": "Open Threads",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "email",
          "name",
          "company",
          "status",
          "account_value",
          "churn_risk_score"
        ],
        "title": "ContactOut"
      },
      "DraftUpdate": {
        "properties": {
          "proposed_content": {
            "type": "string",
            "title": "Proposed Content"
          }
        },
        "type": "object",
        "required": [
          "proposed_content"
        ],
        "title": "DraftUpdate"
      },
      "EmailIn": {
        "properties": {
          "message_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Message Id"
          },
          "sender": {
            "type": "string",
            "format": "email",
            "title": "Sender"
          },
          "subject": {
            "type": "string",
            "title": "Subject",
            "default": ""
          },
          "body": {
            "type": "string",
            "title": "Body",
            "default": ""
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "thread_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Thread Id"
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "sender",
          "timestamp",
          "thread_id"
        ],
        "title": "EmailIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "email_id": {
            "type": "integer",
            "title": "Email Id"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "email_id",
          "duplicate",
          "status",
          "priority_score"
        ],
        "title": "IngestResponse"
      },
      "ReplyRequest": {
        "properties": {
          "body": {
            "type": "string",
            "title": "Body"
          },
          "approved_by": {
            "type": "string",
            "title": "Approved By",
            "default": "user"
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "ReplyRequest"
      },
      "StatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "VIP",
              "Blocked",
              "Active",
              "Churned"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StatusUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}